Apply code review fixes: security, performance, type consistency
Security improvements:
- SEC-008: Add plan node counting and limit validation

Performance optimizations:
- PERF-002: Use spawn_blocking for async storage I/O
- PERF-005: Filter operator avoids copies when 100% selectivity
- PERF-006: Reuse HashMap allocation in group-by aggregation

Type consistency:
- TYPE-001: Change DataBatch series_ids from Vec<u64> to Vec<SeriesId> (u128)
- Update all operators and tests to use consistent SeriesId type
- Remove unnecessary u64/u128 casts

+++ src/query/integration.rs
@@ -125,7 +125,8 @@ impl QueryEngine {
                 rows.push(ResultRow {
                     timestamp: batch.timestamps[i],
                     value: batch.values[i],
-                    series_id: batch.series_ids.as_ref().map(|ids| ids[i] as u128),
+                    // SeriesId is already u128 (TYPE-001), no cast needed
+                    series_id: batch.series_ids.as_ref().map(|ids| ids[i]),
                     tags: HashMap::new(),
                 });
             }
